requirements.txt
py-clob-client
flask
python-dotenv
waitress
